[BEAM-6964] Simplify key context handling in ExecutableStageDoFnOperator
The key handling code in ExecutableStageDoFnOperator was a bit hard to
understand because of the different contexts in which the current key is
set. This simplifies the key setting by always using the state backend to set
the key. An additional guard has been added to ensure that no concurrent access
is performed during key access.

This also adds an additional class for consistent use of key coders.