[BEAM-7029] Add KafkaIO.Read as an external transform
This adds KafkaIO.Read as an external transform and includes a Python
wrapper (ReadFromkafka) for convenience. The transform only returns the
key/value data for a Kafka topic. It does not include the meta data such as
partition id, offset, or timestamp.

By default, the data is returned as `KV<byte[], byte[]>`. Users can supply a
Kafka Deserializer in ReadFromkafka such as LongDeserializer which will infer a
different coder. Only a limited amount of Deserializers are supported.
Alternatively, users can implement their own decoding in the target SDK.